util: fix: notice 'ftell' errors
Coverity observes:

  lib/expr/exeval.c:1397:
  CID 1681735: (#1 of 1): Argument cannot be negative (NEGATIVE_RETURNS)
  7. negative_returns: size is passed to a parameter that cannot be negative.

Despite the wording, this appears to be a true positive. `ftell` can fail for a
variety of reasons but `gv_ftell` was written as if it could not.

Coverity: https://scan5.scan.coverity.com/#/project-view/66357/10724?selectedIssue=1681735
Reported-by: Coverity 2024.12.1
Thanks-to: Giuseppe D'Angelo

+++ plugin/gdk/gvloadimage_gdk.c
@@ -12,6 +12,7 @@
 
 #include <assert.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <util/alloc.h>
@@ -40,12 +41,12 @@ static void gdk_set_mimedata_from_file (cairo_surface_t *image, const char *mime
     if (fp == NULL)
         return;
     fseek (fp, 0, SEEK_END);
-    const size_t len = gv_ftell(fp);
+    const int64_t len = gv_ftell(fp);
     rewind(fp);
     if (len > 0)
-        data = malloc(len);
+        data = malloc((size_t)len);
     if (data) {
-        if (fread(data, len, 1, fp) != 1) {
+        if (fread(data, (size_t)len, 1, fp) != 1) {
             free (data);
             data = NULL;
         }
